app/function_app/requirements.txt
presidio-analyzer
azure-storage-blob
azure-functions

# Testing dependencies
# pytest==7.4.4
# pytest-cov==4.1.0
# coverage==7.4.1
